feat: add Istighfar challenge notifications and streamline stats generation
- Add `sendIstighfarChallengeRank1Notification` to daily stats script to notify users of the daily Istighfar challenge winner
- Remove `roundKey` dependency from Dhikr and Baqiyat challenge notifications, sourcing nicknames directly from the leaderboard entry
- Fix `FirestoreMirror` to use integer increments for counts instead of converting to doubles
- Update `SaloFirebaseMessagingService` to include `FirestoreMirror` dependency when writing FCM tokens

+++ app/src/commonMain/kotlin/tools/mo3ta/salo/data/firebase/FirestoreMirror.kt
@@ -28,7 +28,7 @@ class FirestoreMirror {
                 mapOf(
                     "uid" to uid,
                     "countryCode" to countryCode,
-                    "totalCount" to FieldValue.increment(delta.toDouble()),
+                    "totalCount" to FieldValue.increment(delta),
                     "updatedAt" to FieldValue.serverTimestamp,
                     "nickname" to nickname,
                 ),
@@ -41,7 +41,7 @@ class FirestoreMirror {
             fs.collection(ROUNDS_COLLECTION).document(roundKey)
                 .collection(PLAYERS_SUBCOLLECTION).document(uid)
                 .set(
-                    mapOf("totalExternal" to FieldValue.increment(count.toDouble())),
+                    mapOf("totalExternal" to FieldValue.increment(count)),
                     merge = true,
                 )
         }
